backend/requirements.txt
# Flask Core
Flask>=3.0.0
Flask-CORS>=4.0.0

# Database
Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.23
PyMySQL>=1.1.0
Flask-Migrate>=4.0.5

# Authentication & Security
Flask-JWT-Extended>=4.5.3
python-dotenv>=1.0.0
bcrypt>=4.1.2
Flask-Dance[sqla]>=7.0.0
google-auth>=2.25.2

# Validation
marshmallow>=3.20.0
Flask-Marshmallow>=0.15.0
marshmallow-sqlalchemy>=0.29.0

# Payment Gateway
requests>=2.31.0

# File Processing & Storage
Pillow>=10.1.0

# PDF Generation (Backend)
playwright>=1.40.0
aiofiles>=23.2.1

# Production
gunicorn>=21.2.0